Chnaged price int to long
git-svn-id: http://192.168.2.90/svn/bartsy-server/bartsy-server@386 c3b44e55-a100-42f6-9c86-32dbf579054b

+++ grails-app/domain/bartsy/Ingredients.groovy
@@ -11,7 +11,7 @@ class Ingredients {
     
 	long ingredientId
     String name
-    int price
+    long price
 	String available
 	Date dateCreated
 	Date lastUpdated
